moving utils into separate module
