+++ dynamax/hmm/models/bernoulli_hmm.py
@@ -6,7 +6,7 @@ from dynamax.parameters import ParameterProperties
 from dynamax.hmm.models.abstractions import HMMEmissions, HMM
 from dynamax.hmm.models.initial import StandardHMMInitialState, ParamsStandardHMMInitialState
 from dynamax.hmm.models.transitions import StandardHMMTransitions, ParamsStandardHMMTransitions
-from dynamax.utils import pytree_sum
+from dynamax.utils.utils import pytree_sum
 from jaxtyping import Float, Array
 from typing import NamedTuple, Union
