+++ dynamax/hmm/models/gaussian_hmm.py
@@ -14,7 +14,8 @@ from dynamax.distributions import niw_posterior_update
 from dynamax.hmm.models.abstractions import HMM, HMMEmissions
 from dynamax.hmm.models.initial import StandardHMMInitialState, ParamsStandardHMMInitialState
 from dynamax.hmm.models.transitions import StandardHMMTransitions, ParamsStandardHMMTransitions
-from dynamax.utils import PSDToRealBijector, pytree_sum
+from dynamax.utils.bijectors import RealToPSDBijector
+from dynamax.utils.utils import pytree_sum
 from typing import NamedTuple, Union
 
 
@@ -90,7 +91,7 @@ class GaussianHMMEmissions(HMMEmissions):
             covs=default(emission_covariances, _emission_covs))
         props = ParamsGaussianHMMEmissions(
             means=ParameterProperties(),
-            covs=ParameterProperties(constrainer=tfb.Invert(PSDToRealBijector)))
+            covs=ParameterProperties(constrainer=RealToPSDBijector()))
         return params, props
 
     def collect_suff_stats(self, params, posterior, emissions, inputs=None):
@@ -407,7 +408,7 @@ class SharedCovarianceGaussianHMMEmissions(HMMEmissions):
             cov=default(emission_covariance, _emission_cov))
         props = ParamsSharedCovarianceGaussianHMMEmissions(
             means=ParameterProperties(),
-            cov=ParameterProperties(constrainer=tfb.Invert(PSDToRealBijector)))
+            cov=ParameterProperties(constrainer=RealToPSDBijector()))
         return params, props
 
     def distribution(self, params, state, inputs=None):
