+++ dynamax/hmm/models/linreg_hmm.py
@@ -6,7 +6,8 @@ from dynamax.hmm.models.abstractions import HMM, HMMEmissions
 from dynamax.hmm.models.initial import StandardHMMInitialState, ParamsStandardHMMInitialState
 from dynamax.hmm.models.transitions import StandardHMMTransitions, ParamsStandardHMMTransitions
 from dynamax.parameters import ParameterProperties
-from dynamax.utils import PSDToRealBijector, pytree_sum
+from dynamax.utils.utils import pytree_sum
+from dynamax.utils.bijectors import RealToPSDBijector
 from tensorflow_probability.substrates import jax as tfp
 from typing import NamedTuple, Union
 
@@ -81,7 +82,7 @@ class LinearRegressionHMMEmissions(HMMEmissions):
         props = ParamsLinearRegressionHMMEmissions(
             weights=ParameterProperties(),
             biases=ParameterProperties(),
-            covs=ParameterProperties(constrainer=tfb.Invert(PSDToRealBijector)))
+            covs=ParameterProperties(constrainer=RealToPSDBijector()))
         return params, props
 
     def distribution(self, params, state, inputs):
