+++ dynamax/nonlinear_gaussian_ssm/unscented_kalman_filter_test.py
@@ -3,9 +3,7 @@ import jax.numpy as jnp
 from dynamax.nonlinear_gaussian_ssm.unscented_kalman_filter import unscented_kalman_smoother, UKFHyperParams
 from dynamax.nonlinear_gaussian_ssm.sarkka_lib import ukf, uks
 from dynamax.nonlinear_gaussian_ssm.inference_test import random_nlgssm_args
-
-
-from dynamax.utils import has_tpu
+from dynamax.utils.utils import has_tpu
 
 if has_tpu():
     def allclose(x, y):
